+++ src/model/actions/progression/brand-of-the-elements.ts
@@ -1,15 +1,9 @@
 import { ProgressAction } from '../progress-action';
 import { Simulation } from '../../../simulation/simulation';
 import { Buff } from '../../buff.enum';
+import { CraftingJob } from '../../crafting-job.enum';
 
-import { RecipeElement } from '../../recipe-element';
-
-export abstract class BrandAction extends ProgressAction {
-
-  abstract getBuffedBy(): Buff;
-
-  abstract getElement(): RecipeElement;
-
+export class BrandOfTheElements extends ProgressAction {
   _canBeUsed(simulationState: Simulation, linear?: boolean): boolean {
     return true;
   }
@@ -28,17 +22,17 @@ export abstract class BrandAction extends ProgressAction {
 
   getPotency(simulation: Simulation): number {
     let potency = 100;
-    if (simulation.hasBuff(this.getBuffedBy())) {
+    if (simulation.hasBuff(Buff.NAME_OF_THE_ELEMENTS)) {
       potency += 200 * (1 - simulation.progression / simulation.recipe.progress);
-    }
-    if (simulation.recipe.element) {
-      if (simulation.recipe.element === this.getElement()) {
-        potency *= 2;
-      } else {
-        potency /= 2;
-      }
     }
     return potency;
   }
 
+  getIds(): number[] {
+    return [100331, 100332, 100333, 100334, 100335, 100336, 100337, 100338];
+  }
+
+  getLevelRequirement(): { job: CraftingJob; level: number } {
+    return { job: CraftingJob.ANY, level: 37 };
+  }
 }