+++ CHANGELOG.md
@@ -2,6 +2,68 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.1.0-beta.7](https://github.com/ffxiv-teamcraft/simulator/compare/v1.1.0-beta.6...v1.1.0-beta.7) (2019-06-28)
+
+
+### Features
+
+* implemented reuse buff ([134fddc](https://github.com/ffxiv-teamcraft/simulator/commit/134fddc))
+
+
+
+## [1.1.0-beta.6](https://github.com/ffxiv-teamcraft/simulator/compare/v1.1.0-beta.5...v1.1.0-beta.6) (2019-06-28)
+
+
+### Bug Fixes
+
+* **delicate-synthesis:** fixed an issue making it give 2 stacks instead of only one ([48e96ff](https://github.com/ffxiv-teamcraft/simulator/commit/48e96ff))
+
+
+
+## [1.1.0-beta.5](https://github.com/ffxiv-teamcraft/simulator/compare/v1.1.0-beta.4...v1.1.0-beta.5) (2019-06-28)
+
+
+### Bug Fixes
+
+* fixed an issue with delicate synthesis being unusable ([eabe4f6](https://github.com/ffxiv-teamcraft/simulator/commit/eabe4f6))
+
+
+
+## [1.1.0-beta.4](https://github.com/ffxiv-teamcraft/simulator/compare/v1.1.0-beta.3...v1.1.0-beta.4) (2019-06-28)
+
+
+### Features
+
+* added new skills from the shadowbringers patch ([545d827](https://github.com/ffxiv-teamcraft/simulator/commit/545d827))
+
+
+
+## [1.1.0-beta.3](https://github.com/ffxiv-teamcraft/simulator/compare/v1.1.0-beta.2...v1.1.0-beta.3) (2019-06-27)
+
+
+### Bug Fixes
+
+* fixed an issue with name of elements never being usable ([e432236](https://github.com/ffxiv-teamcraft/simulator/commit/e432236))
+
+
+
+## [1.1.0-beta.2](https://github.com/ffxiv-teamcraft/simulator/compare/v1.1.0-beta.1...v1.1.0-beta.2) (2019-06-26)
+
+
+
+## [1.1.0-beta.1](https://github.com/ffxiv-teamcraft/simulator/compare/v1.1.0-beta.0...v1.1.0-beta.1) (2019-06-26)
+
+
+
+## [1.1.0-beta.0](https://github.com/ffxiv-teamcraft/simulator/compare/v1.0.5...v1.1.0-beta.0) (2019-06-26)
+
+
+### Features
+
+* implemented changes from preliminary 5.0 patch notes ([147192b](https://github.com/ffxiv-teamcraft/simulator/commit/147192b))
+
+
+
 ### [1.0.5](https://github.com/ffxiv-teamcraft/simulator/compare/v1.0.4...v1.0.5) (2019-06-23)